moves getTasks to its own file
